Fix all GCC 6.x warnings

+++ drivers/usb/gadget/f_fs.c
@@ -1550,7 +1550,7 @@ static void functionfs_unbind(struct ffs_data *ffs)
 static int ffs_epfiles_create(struct ffs_data *ffs)
 {
 	struct ffs_epfile *epfile, *epfiles;
-	unsigned i, count;
+	short i, count;
 
 	ENTER();
